fixed users data fetching error
